[add] mypage > 관심 tech 영역 추가
